Add automated testing for foyer tutorials notebook
Previously, when breaking API changes occured to the MoSDeF tools, a lot
of our notebooks would break silently.

Without automated testing, this can be quite difficult to keep track of,
and it is not uncommon for these notebooks to quickly fall out of sync
with the current updates.
This PR introduces a GitHub Action to create a conda environment, and
use the pytest plugin `nbval` to run tests on each notebook. This will
occur daily, for pushes and pull requests.